Update index.md
The default withdraw setting in the Helm Charts is one (1) hour
An advantage of the one hour default setting is assurance that a deployment may not breach the escrow account dramatically. 

This is really bad for any new provider as this will cost alot more and should not be the default ever.
I removed this line  because it is no longer true, Leases close automatically once the balance is finished.